requirements.txt
# --- WEB FRAMEWORK ---
Flask==3.0.0
gunicorn==21.2.0
werkzeug>=3.0.0

# --- DATA SCIENCE & ML CORE ---
numpy==1.24.3
scipy==1.14.1
joblib==1.4.2
scikit-learn==1.6.1
tensorflow==2.13.0
lightgbm==4.6.0
umap-learn==0.5.9.post2

# --- IMAGE PROCESSING ---
pillow>=9.0.0
scikit-image==0.24.0
PyWavelets==1.8.0
opencv-python-headless==4.10.0.84

# --- UTILITIES & API ---
python-dotenv==1.2.1
openai>=1.0.0
httpx==0.27.2
